Move Flight.tsx to Flights directory
